more changes to TiBaseFile and friends to support using streams when reading and writing files generally. some formatting updates in FileProxy, and added a new "length" command in fastdev so we can peak before we stream an entire file across. TIMOB-4612

+++ android/titanium/src/org/appcelerator/titanium/io/TiFile.java
@@ -185,7 +185,7 @@ public class TiFile extends TiBaseFile
 	}
 
 	@Override
-	public double size()
+	public long size()
 	{
 		return file.length();
 	}
